+++ src/autorest-core/lib/parsing/text-utility.ts
